requirements.txt
beautifulsoup4==4.9.1
bs4==0.0.1
certifi==2020.6.20
chardet==3.0.4
python-decouple==3.3
idna==2.10
numpy==1.19.1
pandas==1.1.0
python-dateutil==2.8.1
pytz==2020.1
requests==2.24.0
six==1.15.0
soupsieve==2.0.1
urllib3==1.25.10
pymongo==3.11.0
twilio==6.45.0